feat: add aggressive cache headers
Everything except bootstrap.css and sass.sync.js gets content hashes
and so can be cached indefinitely.

bootstrap.css we never update and are intending to remove, so I did not
create an exception for it.

sass.sync.js is probably never getting updated, but I created an
exception just in case.