fix: update sass and bootstrap headers
The combination of stale-while-revalidate and must-revalidate means that
browsers will keep using their cache until either
max-age + stale-while-revalidate has passed OR max-age has passed and
revalidation fails.

Without must-revalidate, the browser is allowed to re-use the cache
while it fetches the updated version, but that is undesirable in our
case because our users would have to reload the page twice to get
new resources.